Update dependencies and replace old authenticator
Dependencies in composer.json have been updated including a major version bump for saloonphp/saloon and saloonphp/rate-limit-plugin. The authenticator has also been changed to now use Saloon's Http instead of Contracts. This change streamlines the code and aligns with the updates in dependencies. The changes in Saloon version also necessitated an update in composer.lock.