Update dependencies and revise authenticator
The codebase has been updated with new versions of dependencies, notably including a significant version upgrade for 'saloonphp/saloon' and 'saloonphp/rate-limit-plugin'. Additionally, the old authenticator has been replaced with Saloon's Http, discarding the use of Contracts. This adjustment not only cleans up and streamlines the code, but also matches the changes in our updated dependencies. The shift in Saloon version also necessitated corresponding revisions in the composer.lock file.